PostTranslationData.translocation now requires set
The translocation pathway usage has no value associated with them.
Therefore this was refactored to include a set of the pathways used,
not a dict. The JSONSCHEMA and json dumping/loading was edited
to reflect this.

Add setters for elemental_contribution in SubreactionData and
stoichiometry in StoichiometricData

Correct mnm5U element_contribution and update cmo5U element_contribution
and stoichiometry. Cmo5U begins with a conversion to
5-hydroxyuridine, which was unaccounted for an has an
unknown mechanism. Water/protons were added to the stoichiometry
to mass balance the  addition of the hydroxyl group.