Clean up biomass constraints
Add biomass constraint types in one model method as well as conversion
of biomass types to global "biomass"

This requires a "biomass_type" attribute for PosttranslationData to know
what biomass constraint to create as a result of modifications from
SubreactionData. The JSONSCHEMA and json loading/dumping was modified to
reflect this change.